+++ spec/dummy/client/entries/app-kitten.js
@@ -134,6 +134,12 @@ import { RightArrowIcon } from 'kitten/components/icons/right-arrow-icon'
 import { TwitterIcon } from 'kitten/components/icons/twitter-icon'
 import { VisaIcon } from 'kitten/components/icons/visa-icon'
 
+// Illustrations
+import { LinkBoxIllustration } from 'kitten/components/illustrations/link-box-illustration'
+import { LoudspeakerIllustration } from 'kitten/components/illustrations/loudspeaker-illustration'
+import { MixIllustration } from 'kitten/components/illustrations/mix-illustration'
+import { PhoneIllustration } from 'kitten/components/illustrations/phone-illustration'
+
 // Images
 import { KarlImageCropper } from 'kitten/karl/images/image-cropper'
 
@@ -355,6 +361,12 @@ ReactOnRails.register({
   RightArrowIcon,
   TwitterIcon,
   VisaIcon,
+
+  // Illustrations
+  LinkBoxIllustration,
+  LoudspeakerIllustration,
+  MixIllustration,
+  PhoneIllustration,
 
   // Images
   KarlImageCropper,